microservices/image-compression/requirements.txt
fastapi
uvicorn
kafka-python
aiokafka
cassandra-driver
imageio
Pillow